PaginationInfo: convert tests to RTL
